feature: [Entity view] Change id section placement
