+++ apps/ai-dial-admin/src/components/Adapter/View/TabsContent.tsx
@@ -3,20 +3,26 @@
 import { FC } from 'react';
 
 import AdapterModels from '@/src/components/Adapter/ModelsView/AdapterModels';
+import PropertiesTabContent from '@/src/components/EntityTabs/PropertiesTabContent';
 import EntityAudit from '@/src/components/EntityView/Audit/EntityAudit';
+import { DialAdapter } from '@/src/models/dial/adapter';
 import { ApplicationRoute } from '@/src/types/routes';
 import { EntityViewTab } from '@/src/utils/tabs/utils';
-import PropertiesTabContent, { PropertiesProps } from './Properties/TabContent';
+import AdapterProperties from './Properties/Properties';
 
-interface Props extends PropertiesProps {
+interface Props {
   activeTab: EntityViewTab;
+  selectedAdapter: DialAdapter;
+  onChangeAdapter: (adapter: DialAdapter) => void;
 }
 
 const TabsContent: FC<Props> = ({ activeTab, selectedAdapter, onChangeAdapter }) => {
   return (
     <>
       {activeTab === EntityViewTab.Properties && (
-        <PropertiesTabContent selectedAdapter={selectedAdapter} onChangeAdapter={onChangeAdapter} />
+        <PropertiesTabContent entity={selectedAdapter} view={ApplicationRoute.Adapters} id={selectedAdapter.name}>
+          <AdapterProperties entity={selectedAdapter} onChangeAdapter={onChangeAdapter} isEntityImmutable={true} />
+        </PropertiesTabContent>
       )}
       {activeTab === EntityViewTab.Models && (
         <AdapterModels adapter={selectedAdapter} onChangeAdapter={onChangeAdapter} />
