+++ apps/ai-dial-admin/src/components/ApplicationRunners/View/TabsContent.tsx
@@ -4,24 +4,27 @@ import { FC } from 'react';
 
 import AppRunnerApplications from '@/src/components/ApplicationRunners/ConfigurationView/Applications';
 import AppRunnerFeatures from '@/src/components/ApplicationRunners/ConfigurationView/Features';
-import PropertiesTabContent, {
-  PropertiesProps,
-} from '@/src/components/ApplicationRunners/ConfigurationView/TabContent';
 import ApplicationParametersTab from '@/src/components/Applications/ParametersTab/ParametersTab';
 import EntityRoutes from '@/src/components/EntityView/AppRoute/AppRoute';
 import EntityAudit from '@/src/components/EntityView/Audit/EntityAudit';
 import EntityInterceptors from '@/src/components/EntityView/Interceptors/Interceptors';
+import { DialApplicationScheme } from '@/src/models/dial/application';
 import { DialInterceptor } from '@/src/models/dial/interceptor';
 import { DialRole } from '@/src/models/dial/role';
 import { ExportFormat } from '@/src/types/export';
 import { ApplicationRoute } from '@/src/types/routes';
 import { EntityViewTab } from '@/src/utils/tabs/utils';
+import PropertiesTabContent from '@/src/components/EntityTabs/PropertiesTabContent';
+import SchemeProperties from '../ConfigurationView/Properties';
 
-interface Props extends PropertiesProps {
+interface Props {
   activeTab: EntityViewTab;
   selectedFormat: ExportFormat;
   roles: DialRole[];
   interceptors: DialInterceptor[];
+  selectedRunner: DialApplicationScheme;
+  names: string[];
+  onChange: (runner: DialApplicationScheme) => void;
 }
 
 const TabsContent: FC<Props> = ({
@@ -37,7 +40,13 @@ const TabsContent: FC<Props> = ({
     selectedFormat === ExportFormat.ADMIN && (
       <>
         {activeTab === EntityViewTab.Properties && (
-          <PropertiesTabContent names={names} selectedRunner={selectedRunner} onChange={onChange} />
+          <PropertiesTabContent
+            entity={selectedRunner}
+            view={ApplicationRoute.ApplicationRunners}
+            id={selectedRunner.$id}
+          >
+            <SchemeProperties names={names} runner={selectedRunner} isImmutable={true} onChangeRunner={onChange} />
+          </PropertiesTabContent>
         )}
 
         {activeTab === EntityViewTab.Parameters && (
