+++ apps/ai-dial-admin/src/components/InterceptorTemplates/View/TabsContent.tsx
@@ -4,19 +4,29 @@ import { FC } from 'react';
 
 import EntityAudit from '@/src/components/EntityView/Audit/EntityAudit';
 import Interceptors from '@/src/components/InterceptorTemplates/View/Interceptors/Interceptors';
+import { InterceptorTemplate } from '@/src/models/interceptor-template';
 import { ApplicationRoute } from '@/src/types/routes';
 import { EntityViewTab } from '@/src/utils/tabs/utils';
-import PropertiesTabContent, { PropertiesProps } from '../Properties/TabContent';
+import PropertiesTabContent from '../../EntityTabs/PropertiesTabContent';
+import ExtendedProperties from '../Properties/ExtendedProperties';
 
-interface Props extends PropertiesProps {
+interface Props {
   activeTab: EntityViewTab;
+  selectedTemplate: InterceptorTemplate;
+  onChange: (template: InterceptorTemplate) => void;
 }
 
 const TabsContent: FC<Props> = ({ activeTab, selectedTemplate, onChange }) => {
   return (
     <>
       {activeTab === EntityViewTab.Properties && (
-        <PropertiesTabContent selectedTemplate={selectedTemplate} onChange={onChange} />
+        <PropertiesTabContent
+          entity={selectedTemplate}
+          view={ApplicationRoute.InterceptorTemplates}
+          id={selectedTemplate.name}
+        >
+          <ExtendedProperties template={selectedTemplate} onChange={onChange} />
+        </PropertiesTabContent>
       )}
       {activeTab === EntityViewTab.Interceptors && <Interceptors interceptorList={selectedTemplate.interceptors} />}
       {activeTab === EntityViewTab.Audit && (
