+++ apps/ai-dial-admin/src/components/TestSuites/TestCases/Import/ImportFile.tsx
@@ -3,11 +3,11 @@
 import { DialFormPopup, DialLoadFileAreaField, PopupSize } from '@epam/ai-dial-ui-kit';
 import { FC } from 'react';
 
-import { useI18n } from '@/src/locales/client';
 import { BasicI18nKey, ButtonsI18nKey, ImportI18nKey, TestSuitesI18nKey } from '@/src/constants/i18n';
-import { DialLoadFileArea } from '@epam/ai-dial-ui-kit/dist/src/components/LoadFileArea/LoadFileArea';
+import { useI18n } from '@/src/locales/client';
 
 interface Props {
+  selectedTestSuiteId: string;
   isModalOpen: boolean;
   onClose: () => void;
   onApply: () => void;
@@ -17,6 +17,13 @@ const ImportFileModal: FC<Props> = ({ isModalOpen, onClose, onApply }) => {
   const t = useI18n();
 
   const onChangeFile = (files: File[]) => {
+    const body = new FormData();
+
+    body.append('file', files[0] as File);
+
+    // importTestCase(selectedTestSuiteId, body).then((res) => {
+    //   console.log('importTestCase res', res);
+    // });
     // Handle file change
   };
 
@@ -36,7 +43,7 @@ const ImportFileModal: FC<Props> = ({ isModalOpen, onClose, onApply }) => {
         <DialLoadFileAreaField
           elementId="ddd"
           fieldTitle="fff"
-          acceptTypes=".csv"
+          acceptTypes="/"
           emptyTextFirstLine={t(ImportI18nKey.DropAnyFile)}
           emptyTextSecondLine={t(BasicI18nKey.Or)}
           emptyButtonLabel={t(ButtonsI18nKey.Browse)}
